changed file sizes from uint64_t to int64_t
so that -1 could be unknown

+++ s3/s3_transport/include/s3_transport_callbacks.hpp
@@ -102,10 +102,10 @@ namespace irods::experimental::io::s3_transport
             virtual ~callback_for_read_from_s3_base() {};
 
             libs3_types::bucket_context& saved_bucket_context; /* To enable more detailed error messages */
-            uint64_t                     offset;       /* For multiple upload */
-            uint64_t                     content_length;
+            int64_t                      offset;       /* For multiple upload */
+            int64_t                      content_length;
             unsigned int                 thread_identifier;
-            uint64_t                     bytes_read_from_s3;
+            int64_t                      bytes_read_from_s3;
             std::string                  shmem_key;
             time_t                       shared_memory_timeout_in_seconds;
 
@@ -145,7 +145,7 @@ namespace irods::experimental::io::s3_transport
                 cache_fstream.seekp(this->offset);
                 cache_fstream.write(libs3_buffer, libs3_buffer_size);
 
-                auto wrote = static_cast<uint64_t>(cache_fstream.tellp()) - this->offset;
+                auto wrote = static_cast<int64_t>(cache_fstream.tellp()) - this->offset;
                 if (wrote>0) {
                     this->offset += wrote;
                     this->bytes_read_from_s3 += wrote;
@@ -198,7 +198,7 @@ namespace irods::experimental::io::s3_transport
 
                 // writing to buffer
 
-                uint64_t bytes_to_write = this->offset + libs3_buffer_size > output_buffer_size
+                int64_t bytes_to_write = this->offset + libs3_buffer_size > output_buffer_size
                     ? output_buffer_size - this->offset
                     : libs3_buffer_size;
 
@@ -207,14 +207,14 @@ namespace irods::experimental::io::s3_transport
                 this->offset += bytes_to_write;
                 this->bytes_read_from_s3 += bytes_to_write;
 
-                return ((bytes_to_write < static_cast<uint64_t>(libs3_buffer_size)) ?
+                return ((bytes_to_write < static_cast<int64_t>(libs3_buffer_size)) ?
                         S3StatusAbortedByCallback : libs3_types::status_ok);
 
             }
 
             ~callback_for_read_from_s3_to_buffer() {};
 
-            void set_output_buffer_size(uint64_t size)
+            void set_output_buffer_size(int64_t size)
             {
                 output_buffer_size = size;
             }
@@ -227,7 +227,7 @@ namespace irods::experimental::io::s3_transport
         private:
 
             output_char_type *output_buffer;
-            uint64_t         output_buffer_size;
+            int64_t          output_buffer_size;
 
     };
 
@@ -322,11 +322,11 @@ namespace irods::experimental::io::s3_transport
                 std::string                  shmem_key;
                 time_t                       shared_memory_timeout_in_seconds;
 
-                uint64_t                     offset;
-                uint64_t                     content_length;
+                int64_t                      offset;
+                int64_t                      content_length;
                 libs3_types::bucket_context& saved_bucket_context; // To enable more detailed error messages
                 upload_manager&              manager;
-                uint64_t                     bytes_written;
+                int64_t                      bytes_written;
 
                 // Counter incremented each data callback.  Every Nth iteration touch shared memory
                 // so that we know the process didn't die and leave shared memory corrupted
@@ -362,15 +362,15 @@ namespace irods::experimental::io::s3_transport
                     }
 
                     // writing cache file to s3 buffer
-                    uint64_t length_to_read_from_cache = this->content_length - this->bytes_written
-                        > static_cast<uint64_t>(libs3_buffer_size)
-                        ? static_cast<uint64_t>(libs3_buffer_size)
+                    int64_t length_to_read_from_cache = this->content_length - this->bytes_written
+                        > static_cast<int64_t>(libs3_buffer_size)
+                        ? static_cast<int64_t>(libs3_buffer_size)
                         : this->content_length - this->bytes_written;
 
                     cache_fstream.seekg(this->offset);
                     cache_fstream.read(libs3_buffer, length_to_read_from_cache);
 
-                    auto bytes_read_from_cache = static_cast<uint64_t>(cache_fstream.tellg()) - this->offset;
+                    auto bytes_read_from_cache = static_cast<int64_t>(cache_fstream.tellg()) - this->offset;
                     if (bytes_read_from_cache > 0) {
                         this->offset += bytes_read_from_cache;
                         this->bytes_written += bytes_read_from_cache;
@@ -432,7 +432,7 @@ namespace irods::experimental::io::s3_transport
                     }
 
                     // if we've exhausted our current buffer, read the next buffer from the circular_buffer
-                    while (this->offset >= buffer.size()) {
+                    while (this->offset >= static_cast<int64_t>(buffer.size())) {
 
                         upload_page<buffer_type> page;
 
@@ -450,12 +450,12 @@ namespace irods::experimental::io::s3_transport
                         this->offset = 0;
                     }
 
-                    auto remaining_transport_buffer_size = buffer.size() - this->offset;
+                    auto remaining_transport_buffer_size = static_cast<int64_t>(buffer.size()) - this->offset;
 
-                    bool libs3_buffer_larger_than_remaining_transport_buffer = static_cast<uint64_t>(libs3_buffer_size)
+                    bool libs3_buffer_larger_than_remaining_transport_buffer = static_cast<int64_t>(libs3_buffer_size)
                         > remaining_transport_buffer_size;
 
-                    uint64_t length = libs3_buffer_larger_than_remaining_transport_buffer
+                    int64_t length = libs3_buffer_larger_than_remaining_transport_buffer
                         ? remaining_transport_buffer_size
                         : libs3_buffer_size;
 
@@ -638,13 +638,13 @@ namespace irods::experimental::io::s3_transport
                 std::string                  shmem_key;
 
                 unsigned long                sequence;
-                uint64_t                     offset;       // For multiple upload
-                uint64_t                     content_length;
+                int64_t                      offset;       // For multiple upload
+                int64_t                      content_length;
                 libs3_types::bucket_context& saved_bucket_context; // To enable more detailed error messages
 
                 upload_manager&              manager;
 
-                uint64_t                     bytes_written;
+                int64_t                      bytes_written;
 
 
                 // Counter incremented each data callback.  Every Nth iteration touch shared memory
@@ -681,15 +681,15 @@ namespace irods::experimental::io::s3_transport
                     }
 
                     // writing cache file to s3 buffer
-                    uint64_t length_to_read_from_cache = this->content_length - this->bytes_written
-                        > static_cast<uint64_t>(libs3_buffer_size)
-                        ? static_cast<uint64_t>(libs3_buffer_size)
+                    int64_t length_to_read_from_cache = this->content_length - this->bytes_written
+                        > static_cast<int64_t>(libs3_buffer_size)
+                        ? static_cast<int64_t>(libs3_buffer_size)
                         : this->content_length - this->bytes_written;
 
                     cache_fstream.seekg(this->offset);
                     cache_fstream.read(libs3_buffer, length_to_read_from_cache);
 
-                    auto bytes_read_from_cache = static_cast<uint64_t>(cache_fstream.tellg()) - this->offset;
+                    auto bytes_read_from_cache = static_cast<int64_t>(cache_fstream.tellg()) - this->offset;
                     if (bytes_read_from_cache > 0) {
                         this->offset += bytes_read_from_cache;
                         this->bytes_written += bytes_read_from_cache;
@@ -752,7 +752,7 @@ namespace irods::experimental::io::s3_transport
                     }
 
                     // if we've exhausted our current buffer, read the next buffer from the circular_buffer
-                    while (this->offset >= buffer.size()) {
+                    while (this->offset >= static_cast<int64_t>(buffer.size())) {
 
                         upload_page<buffer_type> page;
 
@@ -777,7 +777,7 @@ namespace irods::experimental::io::s3_transport
                     bool libs3_buffer_larger_than_remaining_transport_buffer = static_cast<unsigned long>(libs3_buffer_size)
                         > remaining_transport_buffer_size;
 
-                    uint64_t length = libs3_buffer_larger_than_remaining_transport_buffer
+                    int64_t length = libs3_buffer_larger_than_remaining_transport_buffer
                         ? remaining_transport_buffer_size
                         : libs3_buffer_size;
